Suppress RECS0018 when comparing to zero
RECS0018 : Fix floating point number comparing. Compare a difference with epsilon.
All those zeroes are not from previous calculations, but special values. A valid case of comparing against an epsilon is with Newton crossing, because more often than not, a difference of exactly 0 cannot be reached in floating point arithmetic when applying this iterative algorithm. In that case, an epsilon must be used and I make YOU, the caller, specify it. Otherwise, we would end up with a CPU-bound infinite loop.